Moved component startup to cmd folder inside the component.
- Moving component init code out of the top level package init makes the init predictable.
- Delayed the config parsing for st2reactorcontroller and st2actioncontroller  so that
  import of the component does not automatically cause config parsing.
