+++ st2actioncontroller/bin/action_controller
@@ -1,9 +1,9 @@
 #!/usr/bin/env python2.7
 
 import sys
-from st2actioncontroller import main
+from st2actioncontroller import cmd 
 
 # TODO: Generalize this file so it can be moved to st2common.
 
 if __name__ == '__main__':
-    sys.exit(main())
+    sys.exit(cmd.actioncontroller.main())
