+++ st2actioncontroller/st2actioncontroller/cmd/__init__.py
@@ -1,0 +1,3 @@
+from st2actioncontroller.cmd import actioncontroller
+
+__all__ = ['actioncontroller']
