+++ st2reactor/bin/sensor_container
@@ -5,7 +5,7 @@
 #
 import os
 import sys
-from st2reactor import main
+from st2reactor import cmd 
 
 DEVEL = True
 
@@ -18,4 +18,4 @@ if DEVEL:
     sys.path.append(python_lib_dir)
 
 if __name__ == '__main__':
-    main()
+    cmd.sensorcontainer.main()
