+++ st2reactor/st2reactor/cmd/__init__.py
@@ -1,0 +1,3 @@
+from st2reactor.cmd import sensorcontainer
+
+__all__ = ['sensorcontainer']
