+++ st2reactorcontroller/st2reactorcontroller/cmd/__init__.py
@@ -1,0 +1,3 @@
+from st2reactorcontroller.cmd import reactorcontroller
+
+__all__ = ['reactorcontroller']
